tool: Add flag to avoid catching eval errors
We use `builtins.tryEval` to reduce the chance a check crashes
but that comes at a cost of harder debugging, since the function
does not tell us anything about the error.

Let’s add a CLI flag that removes this check, to allow
observing evaluation errors in their full glory (especially with --show-trace).